backend/requirements.txt
Django==5.2.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.10 